Moved type aliases and syntactic sugar to separate traits
