+++ core/src/main/scala/com/rallyhealth/vapors/core/dsl/factfilter/Dsl.scala
@@ -1,75 +1,16 @@
 package com.rallyhealth.vapors.core.dsl.factfilter
 
-import cats.data.NonEmptyList
 import cats.free.FreeApplicative
 import com.rallyhealth.vapors.core.logic.{Intersect, Union}
 import com.rallyhealth.vapors.core.{algebra, data, dsl}
 
-import scala.language.implicitConversions
 import scala.reflect.ClassTag
 import scala.reflect.runtime.universe.{typeOf, TypeTag}
 
-private[dsl] class Dsl {
+private[dsl] class Dsl extends Types with Syntax {
 
   import algebra._
   import data._
-
-  /**
-    * An alias to this [[dsl]] object, so you can use infix operators and more easily explore
-    * the list of supported expression builders.
-    */
-  final val __ = this
-
-  /**
-    * The root of all expression types.
-    *
-    * @tparam X the type of input
-    * @tparam A the free parameter, used to describe the eventual output of the [[FreeApplicative]] functor
-    */
-  final type AnyExp[X, A] = FreeApplicative[ExpAlg[X, *], A]
-
-  /**
-    * An expression defined over all the given facts, regardless of type, that does not necessarily
-    * terminate into a [[ResultSet]].
-    */
-  final type AnyFactExp[A] = FactsExp[Any, A]
-
-  /**
-    * An expression that terminates into a boolean, used for making a conditional query or filter.
-    *
-    * Useful for defining sub-expressions to [[whereAnyFactHas]], [[whereAllFactsHave]], [[exists]], [[all]], etc.
-    */
-  final type CondExp[X] = AnyExp[X, Boolean]
-
-  /**
-    * A useful alias for building a [[NamedLens]] by passing the identity lens as a starting point to a function.
-    */
-  final type FactLensId[T] = NamedLens.Id[Fact[T]]
-
-  /**
-    * A [[NamedLens]] defined over a [[Fact]].
-    */
-  final type FactLens[T, V] = NamedLens[Fact[T], V]
-
-  /**
-    * Alias for a non-empty list of facts of a certain type.
-    */
-  final type Facts[T] = NonEmptyList[Fact[T]]
-
-  final type TerminalExp[T, V] = AnyExp[V, ResultSet[T]]
-
-  /**
-    * An expression that operates on a non-empty list of facts.
-    *
-    * @tparam T the type of fact
-    * @tparam A the free parameter
-    */
-  type FactsExp[T, A] = AnyExp[Facts[T], A]
-
-  /**
-    * An expression that returns a [[ResultSet]] with metadata and all the facts used to compute the result.
-    */
-  final type TerminalFactsExp[T] = FactsExp[T, ResultSet[T]]
 
   // TODO: Use some cats typeclass instead of iterable?
   def all[F[x] <: IterableOnce[x], T, A](cond: CondExp[T]): CondExp[F[T]] = liftCondExp {
@@ -96,11 +37,6 @@ private[dsl] class Dsl {
   def greaterThanOrEqual[T : Ordering](lowerBound: T): CondExp[T] = liftCondExp {
     ExpAlg.Within[T, Boolean](Window.greaterThanOrEqual(lowerBound), True, False)
   }
-
-  def >[T : Ordering](lowerBound: T): CondExp[T] = greaterThan(lowerBound)
-  def >=[T : Ordering](lowerBound: T): CondExp[T] = greaterThanOrEqual(lowerBound)
-  def <[T : Ordering](upperBound: T): CondExp[T] = lessThan(upperBound)
-  def <=[T : Ordering](upperBound: T): CondExp[T] = lessThanOrEqual(upperBound)
 
   // TODO: Figure out how to use this conditional tertiary expression
 
@@ -138,8 +74,6 @@ private[dsl] class Dsl {
     )
   }
 
-  implicit def logicalOps[T, A](exp: AnyExp[T, A]): LogicalOps[T, A] = new LogicalOps(exp)
-
   def typeNameOf[T : TypeTag]: String = typeOf[T].toString.split('.').dropWhile(_.charAt(0).isLower).mkString(".")
 
   def alwaysTrue[T]: CondExp[T] = liftCondExp(ExpAlg.Pure("True", _ => true))
@@ -156,9 +90,6 @@ private[dsl] class Dsl {
 
   /** @see [[TerminalFactsExp]] */
   private def liftTermExp[T](value: ExpAlg[Facts[T], ResultSet[T]]): TerminalFactsExp[T] = FreeApplicative.lift(value)
-
-  /** @see [[FactsExp]] */
-  private def liftFactsExp[T, A](value: ExpAlg[Facts[T], A]): FactsExp[T, A] = FreeApplicative.lift(value)
 
   /** @see [[CondExp]] */
   private def liftCondExp[X](value: ExpAlg[X, Boolean]): CondExp[X] = FreeApplicative.lift(value)
